backend/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
PyJWT==2.12.1
python-jose==3.3.0
bcrypt==4.1.1
redis==5.0.1
kafka-python==2.0.2
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
click==8.1.7
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2